Implement switching between different views in configure tab
