--- src/__tests__/components/BotBuilder/BotBuilderPages/Configure.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import Configure from '../../../../components/BotBuilder/BotBuilderPages/Configure';
-import { shallow } from 'enzyme';
-
-describe('<Configure />', () => {
-  it('render Configure without crashing', () => {
-    shallow(<Configure />);
-  });
-});
