+++ src/components/BotBuilder/BotBuilder.css
@@ -31,7 +31,7 @@
 	margin: auto;
 }
 .menu-page{
-	margin-top: 20px;
+	margin-top: 10px;
 }
 .preview-wrap{
 	margin-top: 10px;
